requirements.txt
docker==6.1.3
Flask==2.2.5
Flask_Caching==2.0.2
Flask_Cors==3.0.10
python-dotenv==1.0.0
ratelimit==2.2.1
Requests==2.31.0
